pyproject.toml
[project]
name = "nautobot-mcp-server"
version = "0.1.0"
description = "An MCP for performing CRUD ops on Nautobot instances."
authors = [
    {name = "jokelyo",email = "47613996+jokelyo@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
mcp-server = ">=0.1.4,<0.2.0"
pynautobot = "^2.2.1"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.8"
coverage = "^7.10.2"
nautobot = "2.4.10"
pytest = "^8.0.0"
pytest-mock = "^3.12.0"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.23.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.extras]
all = [
]

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | test.py
    | build
    | dist
  )/
  | settings.py
)
'''

[tool.pylint.master]
# Include the pylint_django plugin to avoid spurious warnings about Django patterns
load-plugins = "pylint_django, pylint_nautobot"
ignore = ".venv"

[tool.pylint.basic]
# No docstrings required for private methods (Pylint default),
# or for test_ functions, or for inner Meta classes.
no-docstring-rgx = "^(_|test_|Meta$)"
good-names = ["i", "j", "k", "ex", "Run", "id", "ip", "db", "fp", "fd", "fk"]

[tool.pylint.messages_control]
# Line length is enforced by Black, so pylint doesn't need to check it.
# Pylint and Black disagree about how to format multi-line arrays; Black wins.
disable = """,
    line-too-long,
    R0801,
    """

[tool.pylint.format]
max-line-length = 120

[tool.pylint.miscellaneous]
notes = """,
    FIXME,
    XXX,
    TODO,
"""

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
addopts = "-vv --doctest-modules"

[tool.pylint-nautobot]
supported_nautobot_versions = [
    "2.0.0"
]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = [
    "D",  # pydocstyle
    "F", "E", "W",  # flake8
    "S",  # bandit
    "I",  # isort
]
ignore = [
    # warning: `one-blank-line-before-class` (D203) and `no-blank-line-before-class` (D211) are incompatible.
    "D203", # 1 blank line required before class docstring

    # D212 is enabled by default in google convention, and complains if we have a docstring like:
    # """
    # My docstring is on the line after the opening quotes instead of on the same line as them.
    # """
    # We've discussed and concluded that we consider this to be a valid style choice.
    "D212", # Multi-line docstring summary should start at the first line
    "D213", # Multi-line docstring summary should start at the second line

    # Produces a lot of issues in the current codebase.
    "D401", # First line of docstring should be in imperative mood
    "D407", # Missing dashed underline after section
    "D416", # Section name ends in colon
    "E501", # Line too long
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"development/*.py" = ["F405", "F403"]
"tests/*.py" = ["S101"]  # Allow assert statements in tests

[tool.coverage.run]
disable_warnings = ["already-imported"]
relative_files = true
omit = [
    # Skip Tests
    "*/tests/*",
]


[tool.pymarkdown]
# Seems to be not support for whitelisting rules: https://github.com/jackdewinter/pymarkdown/issues/1396
plugins.md001.enabled = false
plugins.md002.enabled = false
plugins.md003.enabled = false
plugins.md004.enabled = false
plugins.md005.enabled = false
plugins.md006.enabled = false
plugins.md007.enabled = false
plugins.md008.enabled = false
plugins.md009.enabled = false
plugins.md010.enabled = false
plugins.md011.enabled = false
plugins.md012.enabled = false
plugins.md013.enabled = false
plugins.md014.enabled = false
plugins.md015.enabled = false
plugins.md016.enabled = false
plugins.md017.enabled = false
plugins.md018.enabled = false
plugins.md019.enabled = false
plugins.md020.enabled = false
plugins.md021.enabled = false
plugins.md022.enabled = false
plugins.md023.enabled = false
plugins.md024.enabled = false
plugins.md025.enabled = false
plugins.md026.enabled = false
plugins.md027.enabled = false
plugins.md028.enabled = false
plugins.md029.enabled = false
plugins.md030.enabled = false
plugins.md031.enabled = false
plugins.md032.enabled = true # blanks-around-lists
plugins.md033.enabled = false
plugins.md034.enabled = false
plugins.md035.enabled = false
plugins.md036.enabled = false
plugins.md037.enabled = false
plugins.md038.enabled = false
plugins.md039.enabled = false
plugins.md040.enabled = false
plugins.md041.enabled = false
plugins.md042.enabled = false
plugins.md043.enabled = false
plugins.md044.enabled = false
plugins.md045.enabled = false
plugins.md046.enabled = false
plugins.md047.enabled = false
plugins.md048.enabled = false
plugins.md049.enabled = false
plugins.md050.enabled = false
plugins.pml100.enabled = false
plugins.pml101.enabled = true # list-anchored-indent
plugins.pml102.enabled = false
plugins.pml103.enabled = false
